% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrlasso.R
\name{constrlasso}
\alias{constrlasso}
\title{The function constrlasso}
\usage{
constrlasso(
  X,
  y,
  lambda,
  Aeq = NULL,
  beq = NULL,
  A = NULL,
  b = NULL,
  penidx = NULL,
  method = "QP"
)
}
\arguments{
\item{X}{an nxp matrix of p regressors with n observations.}

\item{y}{an nx1 response vector with n observations.}

\item{lambda}{a tuning parameter value for the lasso penalty. Default value is lambda=0.}

\item{Aeq}{a cxp equality constraint matrix, containing c constraints for p regressors.
Default value is Aeq=NULL, no equality constraints.}

\item{beq}{a cx1 equality constraint vector. Default value is beq=NULL, no equality constraints.}

\item{A}{a cxp inequality constraint matrix, containing c constraints for p regressors.
Default value is A=NULL, no inequality constraints.}

\item{b}{a cx1 inequality constraint vector. Default value is b=NULL, no inequality constraints.}

\item{penidx}{a logical px1 vector, indicating which coefficients are to be penalized.
Default value is penidx=NULL and imposes penalty on all p coefficients.}

\item{method}{a character string, the method to be used.
Possible values are "QP" (default) for Quadratic Programming and "CD" for Coordinate Descent.
"CD" uses the glmnet package and only works without equality and inequality constraints.}
}
\value{
betahat a px1 vector of estimated coefficients.

dual_eq equality duals. The function returns an empty vector for no equality constraints.

dual_neq inequality duals. The function returns an empty vector for no inequality constraints.
}
\description{
This function fits a linearly constrained lasso regression,
using a predictor matrix X, a response y and a tuning parameter value lambda.
It results in a vector of coefficient estimates.
The function corresponds to lsq_constrsparsereg of the SparseReg MATLAB-toolbox by Zhou and Gaines,
see \href{http://hua-zhou.github.io/SparseReg/}{the project page}.
Only the Quadratic Programming Algorithm for ENET is implemented.
For more information, see
\href{http://hua-zhou.github.io/media/pdf/GainesKimZhou08CLasso.pdf}{\insertCite{gaines2018algorithms;textual}{constrlasso}}.
}
\section{Details}{

The Constrained Lasso as in \insertCite{gaines2018algorithms;textual}{constrlasso} minimizes
\deqn{0.5||y - X \beta ||^2_2 + \lambda||\beta||_1,}
subject to \eqn{Aeq \beta = beq} and \eqn{A \beta\le b}.
}

\examples{
set.seed(1234)
n <- 200 # number of observations
p <- 150 # number of regressors
real_p <- 50 # number of true predictors
Xmat <- matrix(rnorm(n * p), nrow = n, ncol = p)
yvec <- apply(Xmat[, 1:real_p], 1, sum) + rnorm(n)
results_reg_no_constr <- constrlasso(Xmat, yvec, lambda = 0)

}
\references{
\insertAllCited
}
