% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstrLasso.R
\name{ConstrLassoReg}
\alias{ConstrLassoReg}
\title{The function ConstrLassoReg}
\usage{
ConstrLassoReg(X, y, lambda, Aeq = NULL, beq = NULL, A = NULL,
  b = NULL, penidx = NULL, method = "QP")
}
\arguments{
\item{X}{an nxp matrix of p regressors with n observations.}

\item{y}{an nx1 response vector with n observations.}

\item{lambda}{a tuning parameter value for the lasso penalty. Default value is lambda=0.}

\item{Aeq}{a cxp equality constraint matrix, containing c constraints for p regressors. Default value is Aeq=NULL, no equality constraints.}

\item{beq}{a cx1 equality constraint vector. Default value is beq=NULL, no equality constraints.}

\item{A}{a cxp inequality constraint matrix, containing c constraints for p regressors. Default value is A=NULL, no inequality constraints.}

\item{b}{a cx1 inequality constraint vector. Default value is b=NULL, no inequality constraints.}

\item{penidx}{a logical px1 vector, indicating which coefficients are to be penalized. Default value is penidx=NULL and imposes penalty on all p coefficients.}

\item{method}{a character string, the method to be used. Possible values are "QP" (default) for Quadratic Programming and "CD" for Coordinate Descent.
"CD" uses the glmnet package and only works without equality and inequality constraints.}
}
\value{
betahat a px1 vector of estimated coefficients.

dualEq equality duals. The function returns an empty vector for no equality constraints.

dualIneq inequality duals. The function returns an empty vector for no inequality constraints.
}
\description{
This function fits a linearly constrained lasso regression, using a predictor matrix X, a response y and a tuning parameter value lambda. It results in a vector of coefficient estimates.
The function corresponds to lsq_constrsparsereg of the SparseReg MATLAB-toolbox by Zhou and Gaines, see \href{http://hua-zhou.github.io/SparseReg/}{the project page}.
Only the Quadratic Programming Algorithm for ENET is implemented.
For more information, see \href{http://hua-zhou.github.io/media/pdf/GainesKimZhou08CLasso.pdf}{Gaines et al. (2018)}.
}
\section{Details}{

The Constrained Lasso as in Gaines et al. (2018) minimizes
\deqn{0.5||y - X \beta ||^2_2 + \lambda||\beta||_1,}
subject to \eqn{Aeq \beta = beq} and \eqn{A \beta\le b}.
}

\examples{
library(ConstrLasso)
library(MASS)
set.seed(1234)
n <- 200
p <- 50
Xmat <- matrix(,n,p)
for(i in 1:p){Xmat[,i] <- rnorm(n,runif(1,-3,3),runif(1,1,2))}
betas <- runif(p,-2,2)
nonzeros <- sample(1:p,20,replace=FALSE)
yvec <- Xmat[,nonzeros]\%*\%betas[nonzeros] + rnorm(n,0,2)
classoreg_results <- ConstrLassoReg(Xmat,yvec,lambda=0)
}
